Template types always generate an array of elements
Why: Children were generating an array of dom elements, but others
were generating a single element. Always having an array means that
the appending can be handled in a similar way. Also, you always
have an array of dom elements, even if it is empty. That is easier
on the typescripting.